md_5 is so very tired of picking through diffs, but here is what he has done so far! Note: Still a huge mess.

+++ src/main/java/org/bukkit/craftbukkit/map/CraftMapView.java
@@ -54,7 +54,7 @@ public final class CraftMapView implements MapView {
     }
 
     public World getWorld() {
-        byte dimension = worldMap.map;
+        int dimension = worldMap.map;
         for (World world : Bukkit.getServer().getWorlds()) {
             if (((CraftWorld) world).getHandle().dimension == dimension) {
                 return world;
